% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_create_hs.R
\name{mm_create_hs}
\alias{mm_create_hs}
\title{Create or Add Hierarchical Subject Values in Image Metadata}
\usage{
mm_create_hs(path, value = c())
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}

\item{value}{named character vector specifying the new hierarchical subjects to add.
Each value must have a parent specified as the name, e.g c("Species" = "Vulture").}
}
\description{
Create or Add Hierarchical Subject Values in Image Metadata
}
\examples{

# Define the URL of the image to be downloaded
image_url <- "https://cdn.britannica.com/02/162502-050-FEEA94DE/Vulture.jpg"

# Define the path to save the downloaded image
temp_image_path <- file.path(tempdir(), "image.jpg")

# Download the image from the internet
download.file(url = image_url, destfile = temp_image_path, mode = "wb")

# Get Hierarchical Subject from the image - Before use ti_create_hs()
ti_get_hs(path = temp_image_path) #==> NULL

ti_create_hs(path = temp_image_path, value = c("Species" = "Vulture"))

# Get Hierarchical Subject from the image - Before use ti_create_hs()
ti_get_hs(path = temp_image_path) #==> "Species|Vulture"

}
