% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_get_hs.R
\name{mm_get_hs}
\alias{mm_get_hs}
\title{Retrieve Hierarchical Subject Values from Image Metadata}
\usage{
mm_get_hs(path)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}
}
\value{
A character vector of unique hierarchical subjects if they exist, otherwise NULL.
}
\description{
The function retrieves the values from the \verb{Hierarchical Subject}
field of an image's metadata. It uses exiftool to read the metadata and processes
the results to extract and return the unique hierarchical subjects.
}
\examples{

# Define the URL of the image to be downloaded
image_url <- "https://cdn.britannica.com/02/162502-050-FEEA94DE/Vulture.jpg"

# Define the path to save the downloaded image
temp_image_path <- file.path(tempdir(), "image.jpg")

# Download the image from the internet
download.file(url = image_url, destfile = temp_image_path, mode = "wb")

# Get Hierarchical Subject from the image
ti_get_hs(path = temp_image_path)

}
