% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_remove_hs.R
\name{mm_remove_hs}
\alias{mm_remove_hs}
\title{Remove Specific or All HierarchicalSubject Values from Image Metadata}
\usage{
mm_remove_hs(path, hierarchy = NULL)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}

\item{hierarchy}{A named character vector, e.g c("Species" = "Vulture") specifying the hierarchy to be removed.
If NULL, the entire \verb{Hierarchical Subject} field is removed.}
}
\value{
message indicating image updated
}
\description{
The mm_remove_hs function removes a specific hierarchy from the
\verb{Hierarchical Subject} field in an image's metadata, or it removes the entire
field if no specific hierarchy is provided. It uses exiftool
to manipulate the metadata and ensure that only the desired changes are applied.
}
\examples{

# Define the URL of the image to be downloaded
image_url <- "https://cdn.britannica.com/02/162502-050-FEEA94DE/Vulture.jpg"

# Define the path to save the downloaded image
temp_image_path <- file.path(tempdir(), "image.jpg")

# Download the image from the internet
download.file(url = image_url, destfile = temp_image_path, mode = "wb")

# Get Hierarchical Subject from the image
no_hs <- mm_get_hs(path = temp_image_path)
mm_create_hs(temp_image_path, c("A" = "AB"))
mm_remove_hs(temp_image_path, c("A" = "AB"))

}
