% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_stack_df.R
\name{mm_stack_df}
\alias{mm_stack_df}
\title{Stack a list of data frame}
\usage{
mm_stack_df(df_list)
}
\arguments{
\item{df_list}{list of data frame to be stacked}
}
\value{
data frame
}
\description{
The function takes a list of data frames and stacks them into a
single data frame. It ensures that all columns from the input data frames in the list are
included in the output, filling in missing columns with NA values where necessary.
}
\examples{

x <- data.frame(age = 15, fruit = "Apple", weight = 12)
y <- data.frame(age = 51, fruit = "Tomato")
z <- data.frame(age = 26, fruit = "Lemo", weight = 12, height = 45)
alldf <- list(x,y,z)
mm_stack_df(alldf)
}
