% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_remove_hs.R
\name{ti_remove_hs}
\alias{ti_remove_hs}
\title{Remove Specific or All HierarchicalSubject Values from Image Metadata}
\usage{
ti_remove_hs(path, hierarchy = NULL)
}
\arguments{
\item{path}{A character vector specifying the full path of the image file.}

\item{hierarchy}{A named character vector, e.g c("Species" = "Vulture") specifying the hierarchy to be removed.
If NULL, the entire \verb{Hierarchical Subject} field is removed.}
}
\value{
message indicating image updated
}
\description{
The ti_remove_hs function removes a specific hierarchy from the
\verb{Hierarchical Subject} field in an image's metadata, or it removes the entire
field if no specific hierarchy is provided. It uses exiftool
to manipulate the metadata and ensure that only the desired changes are applied.
}
