% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Update list}
\usage{
update_list(first_list, second_list)
}
\description{
Update an existing list basing on item in second list.
If the element is in the list, append the value from second list
and ensure the values are unique. If the element is not in the list, add it with its value
}
